+++ packages/zefyr/lib/src/widgets/common.dart
@@ -6,10 +6,10 @@ import 'package:flutter/widgets.dart';
 import 'package:notus/notus.dart';
 
 import 'editable_box.dart';
-import 'editable_text.dart';
 import 'horizontal_rule.dart';
 import 'image.dart';
 import 'rich_text.dart';
+import 'scope.dart';
 import 'theme.dart';
 
 /// Raw widget representing a single line of rich text document in Zefyr editor.
@@ -43,42 +43,46 @@ class _RawZefyrLineState extends State<RawZefyrLine> {
 
   @override
   Widget build(BuildContext context) {
-    ensureVisible(context);
+    final scope = ZefyrScope.of(context);
+    if (scope.isEditable) {
+      ensureVisible(context, scope);
+    }
     final theme = ZefyrTheme.of(context);
-    final editable = ZefyrEditableText.of(context);
 
     Widget content;
     if (widget.node.hasEmbed) {
-      content = buildEmbed(context);
+      content = buildEmbed(context, scope);
     } else {
       assert(widget.style != null);
-
-      final text = new EditableRichText(
+      content = ZefyrRichText(
         node: widget.node,
         text: buildText(context),
       );
-      content = new EditableBox(
-        child: text,
+    }
+
+    if (scope.isEditable) {
+      content = EditableBox(
+        child: content,
         node: widget.node,
         layerLink: _link,
-        renderContext: editable.renderContext,
-        showCursor: editable.showCursor,
-        selection: editable.selection,
+        renderContext: scope.renderContext,
+        showCursor: scope.showCursor,
+        selection: scope.selection,
         selectionColor: theme.selectionColor,
+        cursorColor: theme.cursorColor,
       );
+      content = CompositedTransformTarget(link: _link, child: content);
     }
 
-    final result = new CompositedTransformTarget(link: _link, child: content);
     if (widget.padding != null) {
-      return new Padding(padding: widget.padding, child: result);
+      return Padding(padding: widget.padding, child: content);
     }
-    return result;
+    return content;
   }
 
-  void ensureVisible(BuildContext context) {
-    final editable = ZefyrEditableText.of(context);
-    if (editable.selection.isCollapsed &&
-        widget.node.containsOffset(editable.selection.extentOffset)) {
+  void ensureVisible(BuildContext context, ZefyrScope scope) {
+    if (scope.selection.isCollapsed &&
+        widget.node.containsOffset(scope.selection.extentOffset)) {
       WidgetsBinding.instance.addPostFrameCallback((_) {
         bringIntoView(context);
       });
@@ -136,34 +140,14 @@ class _RawZefyrLineState extends State<RawZefyrLine> {
     return result;
   }
 
-  Widget buildEmbed(BuildContext context) {
-    final theme = ZefyrTheme.of(context);
-    final editable = ZefyrEditableText.of(context);
-
+  Widget buildEmbed(BuildContext context, ZefyrScope scope) {
     EmbedNode node = widget.node.children.single;
     EmbedAttribute embed = node.style.get(NotusAttribute.embed);
 
     if (embed.type == EmbedType.horizontalRule) {
-      final hr = new ZefyrHorizontalRule(node: node);
-      return new EditableBox(
-        child: hr,
-        node: widget.node,
-        layerLink: _link,
-        renderContext: editable.renderContext,
-        showCursor: editable.showCursor,
-        selection: editable.selection,
-        selectionColor: theme.selectionColor,
-      );
+      return ZefyrHorizontalRule(node: node);
     } else if (embed.type == EmbedType.image) {
-      return new EditableBox(
-        child: ZefyrImage(node: node, delegate: editable.imageDelegate),
-        node: widget.node,
-        layerLink: _link,
-        renderContext: editable.renderContext,
-        showCursor: editable.showCursor,
-        selection: editable.selection,
-        selectionColor: theme.selectionColor,
-      );
+      return ZefyrImage(node: node, delegate: scope.imageDelegate);
     } else {
       throw new UnimplementedError('Unimplemented embed type ${embed.type}');
     }