+++ packages/zefyr/test/widgets/code_test.dart
@@ -11,7 +11,7 @@ void main() {
   group('$ZefyrCode', () {
     testWidgets('format as code', (tester) async {
       final editor = new EditorSandBox(tester: tester);
-      await editor.tapEditor();
+      await editor.pumpAndTap();
       await editor.tapButtonWithIcon(Icons.code);
 
       BlockNode block = editor.document.root.children.first;