+++ packages/zefyr/test/widgets/editable_text_test.dart
@@ -11,10 +11,22 @@ void main() {
   group('$ZefyrEditableText', () {
     testWidgets('user input', (tester) async {
       final editor = new EditorSandBox(tester: tester);
-      await editor.tapEditor();
+      await editor.pumpAndTap();
       final currentValue = editor.document.toPlainText();
       await enterText(tester, 'Added $currentValue');
       expect(editor.document.toPlainText(), 'Added This House Is A Circus\n');
+    });
+
+    testWidgets('autofocus', (tester) async {
+      final editor = new EditorSandBox(tester: tester, autofocus: true);
+      await editor.pump();
+      expect(editor.focusNode.hasFocus, isTrue);
+    });
+
+    testWidgets('no autofocus', (tester) async {
+      final editor = new EditorSandBox(tester: tester);
+      await editor.pump();
+      expect(editor.focusNode.hasFocus, isFalse);
     });
   });
 }